fix the routing
